+++ sdk/dotnet/Types/Input.cs
@@ -21545,8 +21545,8 @@ namespace Pulumi.Kubernetes.Types.Inputs.Extensions
     }
 
     /// <summary>
-    /// DEPRECATED - extensions/v1beta1/Ingress is deprecated by networking/v1beta1/Ingress and not
-    /// supported by Kubernetes v1.20+ clusters.
+    /// DEPRECATED - extensions/v1beta1/Ingress is deprecated by networking.k8s.io/v1beta1/Ingress
+    /// and not supported by Kubernetes v1.20+ clusters.
     /// 
     /// Ingress is a collection of rules that allow inbound connections to reach the endpoints
     /// defined by a backend. An Ingress can be configured to give services externally-reachable
@@ -29879,7 +29879,7 @@ namespace Pulumi.Kubernetes.Types.Inputs.Storage
     }
 
     /// <summary>
-    /// DEPRECATED - storage/v1beta1/CSINode is deprecated by storage/v1/CSINode.
+    /// DEPRECATED - storage/v1beta1/CSINode is deprecated by storage.k8s.io/v1/CSINode.
     /// 
     /// CSINode holds information about all CSI drivers installed on a node. CSI drivers do not need
     /// to create the CSINode object directly. As long as they use the node-driver-registrar sidecar