+++ tests/integration/go/go_test.go
@@ -38,7 +38,8 @@ func TestGo_YAML(t *testing.T) {
 		//Dependencies: []string{
 		//	"github.com/pulumi/pulumi-kubernetes/sdk/v2",
 		//},
-		Quick: true,
+		Quick:                true,
+		ExpectRefreshChanges: true,
 	})
 }